Addition of extensions functions
A new optional parameter has been added to the manage() function.  The property is called "extensions" which is expected to be an array of strings.  For each string in the array, a property is added to the "supports" object of the MQTT payload with the property name being the name of the array member and the value being boolean "true".  This allows the programmer to add one or more custom extensions to be able to receive incoming commands.